requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
boto3==1.34.0
jinja2==3.1.2
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0
rich==13.7.0
typer==0.9.0
